added json fix
Files that contain pure json data were not being copied over to the destination correctly.  Any kind of file, .json or .map in this case, that contained pure json would only contain [object Object] after being pulled from the site.  This branch contains a way to fix this using a conditional that checks the extension name of the file before converting it into a string.